+++ app_usuario/local_farm/pubspec.yaml
@@ -34,8 +34,6 @@ dependencies:
   flutter_staggered_grid_view: ^0.3.0
   curved_navigation_bar: ^0.3.2
 
-  #
-  timeline_tile: ^1.0.0
 
 dev_dependencies:
   flutter_test:
